% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pages_WPP.R
\name{download_pages_WPP}
\alias{download_pages_WPP}
\title{Download Pages from WPP API}
\usage{
download_pages_WPP(target, headers)
}
\arguments{
\item{target}{A string containing the URL of the API's first page.}

\item{headers}{A list of HTTP headers, typically including authorization.}
}
\value{
A data frame containing the combined data from all pages.
}
\description{
Downloads paginated data from the WPP API and combines it into a single data frame.
}
\examples{
headers <- c("Authorization" = Sys.getenv("WPP_ONU_BEARER"))
base_url <- "https://population.un.org/dataportalapi/api/v1"
target <- paste0(base_url, "/indicators/")
indicators <- download_pages_WPP(target, headers)
print(indicators$name)
}
