% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pages_WPP.R
\name{handle_response}
\alias{handle_response}
\title{Handle API Responses}
\usage{
handle_response(response)
}
\arguments{
\item{response}{An \code{httr} response object.}
}
\value{
The content of the response as a character string.
}
\description{
Validates the API response and extracts its content.
}
\keyword{internal}
